Fix operation of 'resume' keyword
Line 30-31 comments indicate Simulator will resume if 'resume' keyword
is present and True; but the if statement will resume in all cases
unless 'resume' is explicitly set False. Changed default for line 32 get
so that if statement operates as per comments.
The other approach is to alter the comments.
Or of course I may be reading this completely wrong!