Updated wunderfixer documentation to use posix date format and equal signs in options
I noticed that none of the single letter options used equal signs
whereas of the 'word' options --config did but no others. All word
options now have equal signs but left single letter options as is,
putting the equal sign in just looked odd.

The 'usage of this utility is at YOUR RISK' notice was removed from the
usage text on 24 Sep by Matthew. Even though it remains in the utility I
have removed it from the wunderfixer usage text in this commit and will
remove it from wunderfixer as well. Not sure if some sort of standard
note or warning should be included instead.