added missing: Repeater support in "set-transmitter-type" command
As I have to use a repeater with my station for me it is essential to
set the repeater number to my Envoy device.
The —info command is updated to reflect the repeater and the receive
enable status (USETX flags)
The receiver field can be used as Tx period with old VP stations as
they only use the values 0, 1 and 4. But validity is not checked as the
code is primarily for the new VP2 and Vue.