Add @guard decorator to __init__() to ensure we get a weedb exception if our Connection object has an initialisation error
__init() in class Connection originally had a try..except around the
MySQLdb.connect() call to catch any errors and return a weedb exception.
This was removed when the code was restructured at commit
b00989a6326dfd0d7a3d2146bce161eaf19c3f06 and the @guard decorator added
to the class. This decorator was removed at commit
69dc7a808d131d0d341070f9c2f5bb2542e91546. Not sure if this is the
pythonic way to do this or if the old try..exception should be
re-instated  - but this eliminates errors in the test suite test_no_db()
test.