Fix vector average wind speed (vecavg) calculation
The vector average wind speed I see displayed is often higher than the average wind speed, but the triangle inequality implies that the vector sum should always be less than or equal to the straight sum. It looks like weewx is dividing the vector sum components (xsum and ysum) by the time the wind vector was nonzero (dirsumtime), not by the time there is wind data (sumtime, including times when the wind is not blowing). This means that, eg, wind at <0,0> for one minute followed by <20,0> for one minute has a vector average of 20 (and a straight average of 10) while blowing <0.1, 0> then <20,0> has a vector average of 10.05 (and straight average of 10.05). This fix just changes weewx to use sumtime instead of dirsumtime for that calculation.