Fix weeimport with calc_missing to write required history before calc
wee_import writes in tranches, but when calc_missing is enabled,
some calculated values require history that is not available
until the tranche is commited.

This patch adds a new function timespanLimit() to check if records
in the tranche need to be written to the archive before calcMissing
is called for the next imported record.