Added detection for increased spurious values
This is done by setting a default max rain rate of 50 cm/h. This value
can be overwritten in the driver configuration using the setting
'max_rain_rate_cmh'. Whenever the driver calculates  a rain rate higher
than 50 cm/h the current rain value is treated as a spurious value.