Move class Stopper to its own module
Originally, this class was included in test_engine.py. When
test_engine.py was run, it would set up the logger. Then it would run
the engine, which would dynamically load "test_engine.Stopper", which
caused test_engine to be loaded *again*, resulting in the logger getting
set up a 2nd time. This resulted in "ResourceWarning" errors because the
syslog socket connection from the first setup was not closed.