V5.0.0b5. Removed rsa dependency.
The rsa dependency resolves to installing `cryptography`. Some
platforms do not have a `cryptography` wheel, so pip tries to compile
it from source. This requires a Rust compiler, so the compilation
usually fails.

The bad news is that MySQL users will have to install `cryptography`
separately. The good news is that life is easier for all other users.