https://github.com/weewx/weewx/issues/867
Set locale when using internationalized skins

This commit enables a 'locale = xx_XX' configuration for a report.
If the locale is set and available, it will be used for the report,
formatting will be done according to this locale.
If the locale is set and not available, an error message will be logged.
If the locale is set and no 'lang = xx' config exists, 'lang' will
be derived from the locale, even if the
locale is not available.
If the locale is set and a 'lang = xx' config exists, 'lang' be set to this
value and not be derived from the locale.