pyproject.toml: Clarify exclusion of top-level files in wheel
Mark README.md and LICENSE.txt as for sdist only, so that they are not
included in a wheel.  This follows the poetry documentation's guidance
not to include such files:

  https://python-poetry.org/docs/pyproject/#exclude-and-include

specifically

  "When a wheel is installed, its includes are unpacked straight into
  the site-packages directory.  Pay attention to include top level
  files and directories with common names like CHANGELOG.md, LICENSE,
  tests or docs only in sdists and not in wheels."

Resolves installation of these files in wheels and thus in
$prefix/lib/python3.xy/site-packages with poetry-core 1.9.1.  (With
poetry-core 2, the default for include is sdist, so this change has no
effect.)

It seems fine to not include them in wheels, as wheels are essentially
binaries (even if they have foo.py).  But, adding them (someplace) is
separable, as this commit's purpose is to align behavior with older
poetry-core to that of current poetry-core.