Some more changes to dev_requirements.txt
Still remains to be solved:
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
types-requests 2.32.4.20250913 requires urllib3>=2, but you have urllib3 1.26.20 which is incompatible.
tox 4.30.2 requires platformdirs>=4.3.8, but you have platformdirs 3.11.0 which is incompatible.